requirements.txt
fastapi
uvicorn
pydantic
prometheus_client
gunicorn
uvloop
PyYAML
